Hey again~

These are (most) of the optimizations I made in `kleur` that took it from 60k op/s to 90k op/s 🎉 

Some additional items of interest that were not covered in the commit messages:

* I used destructed require on `util.format` because you're already supporting Node 6+ only

* In my testing, I saw no need for `hasOpen` and `hasClose` checks. I thought I was pretty thorough but perhaps you can enlighten me

Most of the wins come from avoiding the spread operator & cutting down on the number of functions being created. The `style()` wrapper was unnecessarily invoked _per_ stack item, rather than once per stack. Also, using `[].concat` made a surprisingly bigger impact than I anticipated. 

### Before

```
# Load time
  ansi-colors: 1.126ms
  chalk: 12.708ms
  clorox: 0.900ms

# All Colors
  ansi-colors x 60,881 ops/sec ±0.60% (91 runs sampled)
  chalk x 7,048 ops/sec ±3.77% (71 runs sampled)
  clorox x 1,168 ops/sec ±3.41% (51 runs sampled)

# Stacked colors
  ansi-colors x 13,706 ops/sec ±0.37% (91 runs sampled)
  chalk x 1,683 ops/sec ±4.20% (73 runs sampled)
  clorox x 448 ops/sec ±4.10% (72 runs sampled)

# Nested colors
  ansi-colors x 28,144 ops/sec ±0.38% (95 runs sampled)
  chalk x 3,437 ops/sec ±4.36% (68 runs sampled)
  clorox x 555 ops/sec ±3.07% (44 runs sampled)
```

### After

```
# Load time
  ansi-colors: 1.037ms
  chalk: 11.317ms
  clorox: 0.869ms

# All Colors
  ansi-colors x 82,339 ops/sec ±0.39% (94 runs sampled)
  chalk x 7,237 ops/sec ±4.09% (70 runs sampled)
  clorox x 1,137 ops/sec ±3.93% (65 runs sampled)

# Stacked colors
  ansi-colors x 26,883 ops/sec ±0.16% (91 runs sampled)
  chalk x 1,733 ops/sec ±4.49% (70 runs sampled)
  clorox x 470 ops/sec ±2.21% (42 runs sampled)

# Nested colors
  ansi-colors x 40,450 ops/sec ±0.16% (95 runs sampled)
  chalk x 3,481 ops/sec ±4.58% (69 runs sampled)
  clorox x 550 ops/sec ±2.48% (42 runs sampled)
```

> All results from Node v8.9.0

---

The load time will be _slightly_ lower if/once #12 is merged; and runtime perf will increase too.

But most notably, I was not able to achieve `kleur`'s full 95k op/s simply because of the "bright" color variants. I suppose it's because the `styles` object is significantly larger, that traversal (or any reference to it) is more expensive.

Dropping the "bright" variants brings my results from 82-84k to 90-92k... in the same vein, choosing _one_ method name for clearing styles would be ideal, as it also affect object size.

Lastly, I see slightly less op/s when using `Reflect` vs `Object`. It's a very cool class that I was not aware of, but it's costing roughly 1k op/s in the "All Colors" bench.