@doowb 2018-07-16
@lukeed thanks again for the PR. I requested some minor code changes that don't seem to make a performance impact but either match my coding style or are, what I think, a little more readable.

@jonschlinkert since you contributed the refactor before this, I'd like any input you have on this. It seems like everything is still working and the tests are passing, but let me know if you foresee any "gotchas".