@lukeed 2018-07-16
Sure thing! Running errands now but I'll apply fixes once I get back. I know my code style can appear a little machine-like 😅