@lukeed 2018-07-16
The "All Colors" now consistently runs at 74-76k op/s for me, which I find to be a significant drop.