@doowb 2018-07-16
We must have entirely different machines and environments. I'm on a Macbook Pro from mid-2013 and when I run in node 10, I get 102k-106k, but if I drop to node 8.9 it's ~62k.

Those numbers are both from your original changes and with the updates I requested. I'll look into this some more later.