@jonschlinkert 2018-07-16
> We must have entirely different machines and environments. I'm on a Macbook Pro from mid-2013 and when I run in node 10, I get 102k-106k, but if I drop to node 8.9 it's ~62k.

Same, when I initially ran the benchmarks to see if kleur was indeed faster, this library still loaded faster (until I changed the order of libs being loaded to make kleur faster - this is partially a node.js caching thing), and kleur was faster in the actual benchmarks themselves, but not by nearly as much as listed in the kleur readme. 

Regardless, to me the bottom line is this:

- these changes are great
- we're happy to have them, and they make the code more performant
- but some minor "optimizations" you're seeing @lukeed (or that we're seeing) are not necessarily optimizations on all versions of node.js.
- I agree that readability is slightly more important than micro-optimizations, but not always. Especially in micro-libs where it's unlikely that contributors will make significant contributions, optimization is better IMHO.