@doowb 2018-07-16
@lukeed if you saw better performance, feel free to revert back to your changes. But fwiw I didn't see those same differences on my box.