@jonschlinkert 2018-07-16
> In my testing, I saw no need for hasOpen and hasClose checks. I thought I was pretty thorough but perhaps you can enlighten me

I just noticed this. These checks were there to prevent the same color from wrapping multiple times when nested, which doesn't matter too much in most cases, but caused some issues when newlines appeared in a string with hard breaks in the terminal. For now, I guess it's fine to remove them, we can just add them back when or if someone creates an issue about it.