@lukeed 2018-07-16
Hey guys,

I'm on a Late 2015 iMac w/ a 4GHz Intel Core i7. I reran this morning & the changes are still there. I understand that `indexOf` ~> `includes` is a micro-opt that rarely matters much (and doesn't at all in newer versions of Node). Ultimately it's up to you guys to decide.

The main optimizations are removal of the spreads & excess function creations &mdash; the above is insignificant really.

@jonschlinkert Ah, thanks for that info. I believe I tested those cases pretty well on my end, but not claiming 100% confidence here. Also forgot to play with the ordering on the load bench; that escaped my mind 😅 I just followed the order listed. On Node 8.x, `ansi-colors` does indeed load faster, but node on Node 10.x 😆Node can be funny

```
Node 8.x
---
chalk: 8.854ms
clorox: 0.466ms
kleur: 0.830ms
ansi-colors: 0.691ms

Node 10.x
---
chalk: 9.030ms
clorox: 0.509ms
kleur: 0.576ms
ansi-colors: 0.688ms
```
> Doesn't matter

Also, I don't believe that your claiming this but I'd like to mention it anyway: I wasn't fudging the numbers on kleur results, nor here. That would take too much effort & frankly I don't care enough. Benchmarks are always included in my "fast" modules so that results can be reproduced on _your own_ machine &mdash; I believe you guys commit benchmarks for the same reason~!