@doowb 2018-07-16
@lukeed thanks for the additional information, it's interesting to see how the different environments impact node.

I'll merge this in how it is now. I also added you as a collaborator so feel free to add any other optimizations you feel are necessary.

Thanks again!