@jonschlinkert 2018-07-16
> Also, I don't believe that you're claiming this but I'd like to mention it anyway: I wasn't fudging the numbers on kleur results, nor here.

no no, never. I wasn't implying that at all, it never crossed my mind. Glad you mentioned something to clarify.